Revert "Revert "Set total color dtype to np double""
This reverts commit 1b3756444e2a53e5a31e9f06075c5dd640884270.

+++ skimage/future/graph/rag.py
@@ -359,7 +359,8 @@ def rag_mean_color(image, labels, connectivity=2, mode='distance',
     for n in graph:
         graph.nodes[n].update({'labels': [n],
                                'pixel count': 0,
-                               'total color': np.zeros(img_channels)})
+                               'total color': np.zeros(img_channels,
+                                                      dtype=np.double)})
 
     for index in np.ndindex(labels.shape):
         current = labels[index]
